fix:서스펜스로 페이지 다 감싼거 테스트
